requirements.txt
flask
flask-cors
python-dotenv
insightface
onnxruntime
opencv-python
numpy
supabase
SQLAlchemy
PyMySQL
firebase-admin
cryptography>=42,<45
celery
redis
gunicorn